pyproject.toml
[project]
name = "dholuo-tts"
version = "0.1.0"
description = "🇰🇪 Dholuo Context-Aware Text-to-Speech (TTS) System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=0.26.0",
    "coqui-tts==0.22.1",
    "datasets>=4.5.0",
    "evaluate>=0.4.6",
    "librosa>=0.11.0",
    "pandas>=2.3.3",
    "pyarrow>=23.0.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "seqeval>=1.2.2",
    "soundfile>=0.13.1",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "transformers>=4.38.0,<4.42.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "redis>=5.2.0",
    "python-multipart>=0.0.12",
    "pydantic-settings>=2.6.0",
]

demo = [
    "gradio>=5.0.0",
    "matplotlib>=3.9.0",
    "librosa>=0.11.0",
    "plotly>=5.24.0",
    "pillow>=11.0.0",
    "huggingface-hub>=0.23.2",
]

[tool.setuptools]
packages = ["demo", "api"]
py-modules = ["tagger", "phonemizer", "utils", "main"]
